Fixed registartion

+++ service/src/main/java/by/xenon28082/shop/service/ReserveService.java
@@ -7,6 +7,6 @@ import java.util.List;
 
 public interface ReserveService {
     boolean reserveProduct(Reserve reserve) throws SQLException;
-    List<Reserve> getReservations(int userId) throws SQLException;
+    List<Reserve> getReservations(long userId) throws SQLException;
     boolean deleteReservation(long userId, long orderId);
 }
